Startup:  * Display an image at startup to make the user wait  * Most of operations at startup are now made from another thread, except the loading of textures, which can only be made from the main thread

+++ trunk/src/ta3d/src/gfx/gfx.toolkit.cpp
@@ -567,7 +567,7 @@ namespace TA3D
 		r *= r;
 		for (int sy = -y ; sy <= y ; sy++)
 		{
-			int dx = trunc(sqrtf(r - sy * sy));
+			int dx = (int)trunc(sqrtf(r - sy * sy));
 			int ax = x - dx;
 			int bx = x + dx;
 			for (int sx = ax ; sx <= bx ; sx++)
